W-16562974: Move CompositeRoutingException to privileged package
(cherry picked from commit 9b4ec2e8375cf46583357fc38696ce68a66eef90)
