+++ core/src/main/java/org/mule/runtime/core/internal/exception/ErrorTypeLocatorFactory.java
@@ -41,7 +41,6 @@ import org.mule.runtime.api.security.SecurityException;
 import org.mule.runtime.api.security.ServerSecurityException;
 import org.mule.runtime.api.streaming.exception.StreamingBufferSizeExceededException;
 import org.mule.runtime.api.tx.TransactionException;
-import org.mule.runtime.ast.api.error.ErrorTypeBuilder;
 import org.mule.runtime.core.api.exception.ExceptionMapper;
 import org.mule.runtime.core.api.expression.ExpressionRuntimeException;
 import org.mule.runtime.core.api.retry.policy.RetryPolicyExhaustedException;
@@ -49,7 +48,7 @@ import org.mule.runtime.core.api.transformer.MessageTransformerException;
 import org.mule.runtime.core.api.transformer.TransformerException;
 import org.mule.runtime.core.internal.construct.FlowBackPressureException;
 import org.mule.runtime.core.internal.event.EventContextDeepNestingException;
-import org.mule.runtime.core.internal.routing.result.CompositeRoutingException;
+import org.mule.runtime.core.privileged.routing.CompositeRoutingException;
 import org.mule.runtime.core.internal.routing.result.RoutingException;
 import org.mule.runtime.core.internal.routing.split.DuplicateMessageException;
 import org.mule.runtime.core.privileged.exception.ErrorTypeLocator;
