+++ modules/core-components/src/main/java/org/mule/runtime/core/internal/routing/ForkJoinStrategyFactory.java
@@ -10,7 +10,7 @@ import org.mule.runtime.api.message.ErrorType;
 import org.mule.runtime.api.metadata.DataType;
 import org.mule.runtime.api.scheduler.Scheduler;
 import org.mule.runtime.core.api.processor.strategy.ProcessingStrategy;
-import org.mule.runtime.core.internal.routing.result.CompositeRoutingException;
+import org.mule.runtime.core.privileged.routing.CompositeRoutingException;
 
 
 public interface ForkJoinStrategyFactory {
